Add initial test exercising the XFS metric split
Test qa/645 is added to check the logrewrite config for
the xfs.* and quota.* hierarchies.  Several issues have
been uncovered in the code as a result, and fixed here.

This is an interim commit as there's outstanding issues
with handling the filesys indom - seeking help on that
one from kenj, the rewrite rules don't seem to do what
I want here.