Update the pcp(1) command to be able to report using pmcd.hostname
The pmlogger and pcp commands are the only clients using the timezone
metric from pmcd.  This commit updates the latter to match pmloggers
use of pmcd.hostname as well (i.e. only if available - so recent pmcd
only, else continues to fall back to pmhostname(1) output).

Furthermore, the pcp(1) command was found to not be using the timezone
from the log label (although hostname is extracted from there).  There
appears to be no good reason for this, and on inspection it is usually
reporting "unknown" for archive timezones, since pmcd.timezone is not
usually logged (since its already in the archive label).  Fix this, by
extracting the timezone along with the hostname from the archive label
and then making use of that when reporting archive data.

Finally, update the pcp command pmlogconf file as there was never need
to be logging pmcd.timezone at all, we had it in the label all along.