Make pmcd.hostname metric match pmhostname(1)
Test qa/054 is failing, ultimately due to a failure to
fully resolve the local host name for the pmcd.hostname
metric.

On closer inspection it turns out pmcd.pmlogger.hostname
is also attempting to resolve the host name, and is also
doing it slightly differently to pmhostname(1).  Make em
share code, ensure the hostname is evaluated once on any
fetch (but do not cache forever in case it changes), and
make both use the libpcp networking APIs - which matches
what pmhostname does.

And test qa/054 is passing once again.