pmdagfs2: misc changes and alter glock_array size to PMDA_CACHE_SIZE_ACTIVE
Made changes to the array size calculation in lock_time.c to allocate
only the number of active entries in the cache. An issue was found during
testing where the allocated array size would be inflated because of
PMDA_CACHE_SIZE being used (culled but not yet reclaimed entries in the size
count) instead of PMDA_CACHE_SIZE_ACTIVE resulting in an ever increasing
allocation size for the array.

Extra check introduced in PMDA_CACHE_WALK to ensure return of only active
entries as glock data from the current call should be returned.

Made reductions to help text for gfs2.glock_lock_time metrics as they
were a just a bit lengthy.

Small amendment to 655.out to match an extra line of output that was
not getting caught whilst checking qa/655.