pmlogrewrite - add support for indom replication
A new DUPLICATE option in the indom rewriting rules allows an
instance domain to be copied and the domain number to be rewritten.
Without the DUPLICATE option the original semantics of rewriting
an existing instance domain remains unchanged.

This change allows metrics with a common instance domain to be split
into two PMDAs, as happened in the recent linux -> linux + linux_xfs
split.