Fix the build after recent infiniband spec changes
For those of us with infiniband enabled in the build, recent
merge of spec.in changes introduced a "infinband" typo.  We
do not actually need the spec snippet in question which adds
a conditional "touch".  Its OK to create this list of files
for infiniband (for a non-IB build, has just the man page?);
we wont be using the file anyway as the %files section has a
guard around it.