Remove unused global variables from pmdaproc and pmdaxfs code.
It is no longer valid to run the Linux pmdaproc as a DSO PMDA,
it performs changes to the process euid/egid which are not at
all appropriate for pmcd, which no longer runs in a privileged
mode that can make such changes anyway.  As a result, we can
now remove a global and make some more functions static.

Similarly, pmdaxfs must nowadays run as root, and as a result
must run as a daemon (thats the whole point).  Further things
can be made static here too - indomtab, metrictab, functions -
make it so.