Fix up logconf sar file issues found by test qa/366
The recently added sar pmlogconf template contained many old
metrics that were IRIX-specific, and so test 366 complained.
Rework this file based on the sar2pcp implementation, which
has decent coverage for a sar-pcp metric set.  Split this in
two files, one for log-once (sar-summary), and the rest with
default-interval logging.