Fix build issue on Mac OS X with static probes enabled
Using dtrace as the toolchain, rather than systemtap, shows
a couple of differences in the build we need to deal with -
the probe keyword cannot be used as a provider name and the
building of the object file step (-G to dtrace wrapper) is
not required.

Finally, this splits the initial static probe into two, one
for PDU xmit/recv, and the other for everything else.