Updates to Stans new pmcd static probe test qa/749
Instead of "fake-passing" in the case of missing systemtap,
use _notrun - this gives us insight into which test hosts
are actually running the test, and which not.  Use pmconfig
mechanism to test for static probe support in the installed
pcp under test (not necessarily the same as dev branch).
Switch from hard-coded Red Hat / Fedora paths to pmcd over
to using $PCP_BINADM_DIR, and create the stap script in a
temporary file to allow ease of debugging.

Run stap directly, not in a sub-shell.  Finally, add a few
other PDU types into the mix (pmLookupDesc, pmLookupText),
just to mix things up a bit.