Demonstration programs for the MMV API
Add some demo programs, installed to $PCP_DEMOS_DIR/mmv,
which will be referenced in the PCP Programmers Guide as
the MMV PMDA and client library are being explained.

Add a Makefile and README file into the installed set as
well, to help out anyone trying to figure out this demo.