Fix heap corrption problem in pmcd.
Noticed while running qa test 244. In ResetBadHosts(), nBadHosts
and szBadHosts w were reset to 0, but badHost itself was not set to
NULL, even though it was freed. The address was then given out
on a subequent call to malloc(). Then, when AddBadHost() was called
the same (already reallocated) address was passed to realloc() and
the chaos ensued.

This fix resets badHost to NULL after it is freed.