Keep "golden" copies of each of the books in the tree
The toolchains for generating the books are many and varied,
and unfortunately each seems to have its own set of issues.
Some appear to be relatively new, and/or not support all of
the platforms we'd like to build on.  Others, seem to have
poor attitudes towards back-compat of their own formats and/
or arguments, and are very difficult to make work for all of
the platforms we'd like to build on (ie with different tool
versions).

All of this leads to a lack of trust in book toolchains, but
we do want the books in all pcp-doc releases.  So, commit a
known good version of each book into the tree, and install
that (a bit like the autoconf/configure situation I guess).