Move extended pmhostname diagnostics into libpcp
Some recent additions to help diagnose hostname lookup
issues were exposing deep internals of libpcp networking
APIs outside of libpcp, and in a way that would not work
for certain configure/build combos.  Move these inside
the libpcp APIs that were being instrumented, resolving
the latter issue.  Extra diagnostics for the nspr versions
is left as an exercise for the reader, for now.