Revert part of the unix: ACLs changes related to pmlogconf.sh
The private pmlogger/pmlc does not (yet) know how to distinguish
localhost from af_unix - none of the security/credentials work is
in place there yet, so we must not generate a config using newer
af_unix-capable syntax there at this stage.  This was the cause of
some of the (scarier looking) QA fallout - easily fixed, happily.