Take POSIXLY_CORRECT out of the PMDA environment once used
PMCD sets POSIXLY_CORRECT in the environment for PMDAs to have
that getopt(3) behaviour (for some reason, lost in time a bit).
This is OK, but we should not enforce this on poor unsuspecting
PMDA authors who may be running tools that rely on this env var
not being set.

So, at the end of the common PMDA library getopt handling, we
take it back out of the environment for the PMDA process and we
continue on our merry path.  Nothing can be done for DSO PMDAs,
but this so far has only burned one Perl PMDA (Samba) so that's
possibly not a problem in practice.