Server side filtering of process instance domain (threads/cgroups)
The Linux proc PMDA exports the set of all processes (incl threads)
as its primary instance domain, for all metrics that have per-process
metrics.

This commit provides a mechanism for either per-client or system wide
filtering of that set (process instance domain).  A series of metrics
are added allowing finer-grained filtering of this set, such that all
processes, or all-processes-and-all-threads, or all-processes-within-
a-cgroup, or all processes-and-all-threads-within-a-cgroup can be
requested instead.  This serves to limit the data that needs to be
sent to a client, or stored on disk (or sampled in the kernel, in the
case of cgroups), and provides a convenient way to allow clients to
become cgroup-aware (in terms of the process metrics anyway).

See "pminfo -T proc.control" for further details on the different ways
client tools can make use of this mechanism.