Treat a slash as equivalent to string terminator for hostspec
It was observed that we'll attempt to connect to a host name
that is not what the user expects with a hostspec along the
lines of pcps://foo.com/ as a result of the trailing slash.
Instead of treating this as part of the host name, just end
parsing at that point and ignore it.  Down the track we may
want to allow metric names after this slash, perhaps, but we
can safely ignore it for now.