Remove newly added proc.control.all.cgroups metric
On review, we decided proc.control.all.cgroups might be abused,
and its not likely to be a useful system-wide setting anyway.
Add a command-line option to allow pmdaproc to default into non
thread reporting (process indom contains no threads).  Defaults
to the existing behaviour though, which is to show threads too.