Resolve static probe build issue for exotic Debian platforms
The Debian builds for Hurd and kFreeBSD are (not surprisingly)
not enjoying the presence of static probe support in the build.
They evidently have found sdt.h somewhere (?) but the lack of a
dtrace build tool means the build fails.  We tighten configure
checking in the area of /usr/bin/dtrace, to ensure enable-probes
is disabled completely if it is absent.