Add proc.psinfo metrics for cgroups and security labels
Extract Linux per-process metrics from procfs for each running
process.  Observed that many, many processes have the same value
for these (string) metrics, so optimise the memory used so that
only a single copy of each different value is ever held in-core.
We achieve this through (ab)use of the pmdaCache API for indoms,
using a pseudo-indom to hold a string dictionary, and associate
the (inst) identifier with each process rather than the string.

The string format for each of these metrics is made to match the
format that ps(1) reports - for security labels there is no need
to tweak the kernel value, for cgroups some string manipulation
is performed (via converting the multi-line string to a one-line
semi-colon separated string without the cgroup-hierarchy-number).

Tests 723 & 724 are added to exercise the extraction, comparing
the pmdaproc values to those obtained from reading directly the
procfs files.