pmie bug - archive, interp mode, mark record, ...
Found investigating a pmie core dump on a specific archive.

Problem triggered by an arithmetic expression with two metrics over
the same instance domain, one with INSTANT semantics, the other with
DISCRETE semantics, and an archive with <mark> records and in one of
the periods between <mark> records, the DISCRETE metric had defined
values, but the INSTANT one did not (although it did have values
earlier and later in the archive).

An loop in the binary.sk template was protected by a guard that was
not strong enough in this case ... fixed and now the expression is
undefined in these circumstances as it should be.

Very specific QA case added.