Resolve a pmcollectl divide-by-zero picked up by qa/709
Now that pmcollectl is defaulting to local: connections, it
seems we are much more likely to observe a zero delta in the
network in/out packet values.  This is causing a div-by-zero
python stack trace in pmcollectl.

Resolve this by taking care to ensure the divisor is never
zero.  Also fix a typo in the suppression of loopback for the
network.interface.out metric sum and add a comment suggesting
pmDelProfile would be a better option for tackling this class
of problem (currently works only because lo is presented as
the first entry in /proc/net/dev by the kernel - but actually
multipe loopback devices can be configured, and there may not
be a guarantee that lo0 is always first anyway).