Allow opt-out with the default-enabled xfs and proc PMDAs
Nowadays we run pmdaproc and pmdaxfs separately to the main
kernel PMDA on Linux, and under the root user account.  As
some people may not wish to run additional daemons as root,
provide a Remove script for each.  Of course, someone will
want to undo such an action, at some point, so we'll need to
have Install scripts too.  These need "root" PMNS files to
ensure a successful install with all the bits in the right
places, with the right domain numbers and so on.