When in thread-on mode, make pmdaproc values use /proc/PID/task/NNN
Stan noticed that the threaded cpu utime/stime values were incorrect
and we were simply exporting the same values for the parent as all of
the children.  Turned out this was because we were not looking below
/proc/PID/task/NNN for the threaded values but simply /proc/NNN which
is the group-leader stats only.

Since the thread values have been wrong for awhile, now appears to be
a good time to change the default threaded state to off, and export
only the (always been correct) process values.  This also reduces the
default fetch load for the proc indom, while still giving good data.
The command line runtime-store-able options exist to switch to the now
correct threaded values.  Since tools like ps, top and atop default to
no-threads, this seems like the best route to take.