Functional network.interface speed/duplex metrics on older kernels
Permissions issues on the SIOCETHTOOL ioctl have been causing the
Linux kernel PMDA to see EPERM for older (e.g. pre-3 kernels), as
a result of overly restrictive controls on this read-only ioctl.
This was later rectified, but is still a problem for us since we
do not run pmdalinux as root.

It turns out the information we're after is available through the
sysfs filesystem, so this change adds a fallback to using sysfs if
that ETHTOOL ioctl fails.  Turns out its easier than I thought to
find the information we're after in sysfs - symlinks exist to help
point us the right way, below /sys/class/net.  The prospect of an
entire /sys/devices tree scan was giving me a headache, so pretty
pleased to have stumbled across that method!