pmNewContext: plug memory leak on failed connections
The __pmContext struct that's malloc'd anew early on during
pmNewContext needs to be fred within the FAILED: path, lest we seek to
sink the good ship HMS Server with a memory leak.

The problem was initially found when "pmie -h NOSUCHHOST" goes into a
frenzy, retrying context creation every few seconds.  (Why pmie was
started with -h NOSUCHHOST is left to another surprising and
delightful tale.)