systemd pmda: add uid/gid-filtering mode
As a confidentiality matter, we should not expose system log
information to any Tom, Dick, and Harry.  journalctl has an
filesystem-ACL mechanism for restricting visibility of logs to certain
users/groups, but short of running root & setuid'ing back & forth for
clients, we can't rely on that.

So, we implement a baby ACL ourselves, using the event-queue filtering
logic for each client (authenticated), passing to each client only
those journal entries that match _UID=$uid -or- _GID=$gid -or- have
$gid as one journalctl groups adm/wheel/systemd-journal.  The effect
is not quite the same (since we disregard auxiliary group
memberships), but c'est la view.  We err on the side of restricting
info (to non-root users).

Unauthenticated contexts, by default, get no soup.  A new non-default
command line option -f is available to signify that there does exist
such a thing as free lunch, and thereby to serve journal entries to
anyone who asks.