Make default crontab install cooperate with runlevel settings
We found local pmie and pmlogger processes enabled everywhere,
now that the check crontab scripts are always installed.  Bit
of an oversight, it resulted from service state always being
ignored.  This change adds an option to the check scripts such
that they ensure the system runlevel state is honoured when we
run the check scripts (either stopping or starting processes).