systemd pmda: correct handling multiple concurrent clients
We were using return values of pmdaEventNewClient (which BTW are
undocumented!) as though it were a queue number, whereas we actually
had a good number already as a global variable.

While at it, reduce the systemd poll fallback timeout from 2 to 60
seconds, with the RHBZ979487 workaround in effect, plus prior changes
that preclude us from being flooded with pent-up journal entries.