Make the installed pmie example files consistently built
Indeterminism in the build (and/or build hosts) was causing
different pmie examples files to be generated depending on
the phase of the moon.  Clamp down on this, and ensure that
consistent example pmie files are produced across builds.
In the process, noticed the copyright-checker script from
eons ago went crazy on these (partial) files, resulting in
the built configs having copyright notices repeated many
times throughout the final generated file - fixed!