pmdagfs2: Minor update and coverity defect fixes
As suggested by Nathan have reduced the value of GLOCK_ARRAY_CAPACITY to
2048 (from 100000), this should reduce the chance of malloc failing whilst
keeping the number of malloc's performed on a system under heavy load to an
acceptable level. Renamed gfs2_ftrace(ch) to ftrace(ch) and re-factored the
change throughout the rest of the code.

Fixed outstanding CHECKED_RETURN and RESOURCE_LEAK coverity defects for
pmdagfs2 in ftrace.c and pmda.c respectively and rechecked.

Left NO_EFFECT defect (the only defect left showing for pmdagfs2 on my runs)
in pmda.c because it is intentional.