Add the jsdemos into the source tarball
Andy on IRC pointed out the jsdemos were not in the source tgz.
Since we've not tackled better build/git integration yet, this
is still a bit painful (esp. for deep directory trees).  Make
a checked-in tarball of the jquery-ui-themes, and expand it in
the build, rather than having to deal with such a deep tree.