Implement a Journal Block Device v2 PMDA (ext3,ext4,ocfs2 journal)
A new PMDA exporting the journal statistics from JDB2 procfs entries.
The kernel interface exports a series of mostly-averaged statistics,
which is unfortunate - over time, this becomes smoothed such that new
spikes are difficult to discern.  An unrolling of the averages is also
provided (jdb2.transaction.total.*) which uses the tr_tid to attempt
conversion back to the original counter values.

This is likely to be affected by loss of precision, so some discussion
with the filesystem folks who originally showed interest here has been
had, suggesting those values be exported directly from the kernel with
no averaging.  For now, we'll just have to make do.

Test 725 is added to exercise the new functionality.  Test 273 was not
deterministic when new DSO PMDAs are present in pmcd.conf; this commit
resolves that as well.