Update tests qa/29[45] to cope with changed pmie hostname handling
We recently changed pmie to use pmGetContextHostName in order that
it can better grok the remote hostname.  For hosts, we establish a
temporary context early on to extract the correct hostname for log
file reporting, the stats file, and so on.  This causes QA tests to
fail if they are counting the number of pmcd connects/disconnects,
as they are now off-by-one.

Further, in the default case (i.e. no host/archive options) should
pmie startup with no pmcd running, its handling of hostname lookup
has been improved so that it falls back to using a local-context.
This is important because pmie can be very long-lived and we would
really like it to get a more sensible host name than "?" if at all
possible.  Once its (re)connects a context when pmcd finally comes
up, it may run for a very long time - not such an unusual scenario.