Try harder in pmie to come up with a hostname for local connections
In the default startup case (i.e. no host/archive options) should
pmie startup with no pmcd running, its handling of hostname lookup
has been improved so that it falls back to using a local-context.
This is important because pmie can be very long-lived and we would
really like it to get a more sensible host name than "?" if at all
possible.  Once it establishes a context when pmcd eventually comes
up, it may run for a very long time and cannot nonsense forever -
not such an unusual scenario.  Possibly pmie should really be doing
the pmcd.hostname check whenever it reconnects.

While there, remove an unused variable in pmie initArchive() code,
and make dfltConn local to the argument parsing code instead of a
global.