Rethink the conflicting-Avahi-service-registry-attempted message
We've recently started to generate a warning message when the pmcd
Avahi service needs to be shifted a slot, as there's one already
registered.  This is a normal scenario in the case where more than
one pmcd port has been requested (back-compat mode) so we should
not warn here.  In most cases where this would happen (ie someone
registers "PCP pmcd" as a service), its extremely like it was us.

Its also highly problematic because of where the message is issued,
has to be just prior to opening log files which means it goes into
the ether anyway, usually, rather than being logged.  Made this as
a diagnostic info message guarded by the discovery pmDebug flag as
a result.