configure.in - fix HAVE_CONST_DIRENT test
At some point either the gcc pixies messed with compile-time checking
or the pcp pixies changed the compilation flags so gcc does insane
compile-time checking.

The sample C code was generating an error because the first and second
arguments to scandir() were NULL.  Don't even start to think about a
parallel universe in which this is ever a good idea!

The consequence was that HAVE_CONST_DIRENT is never set, even when
the scandir() callback routine should have a const struct dirent *
argument, so we assume struct dirent * and end up with compilation
errors on some platforms.