Ensure the correct context is used for several libqmc calls
Chandana reported that opening an archive in pmchart, after
first starting it with host contexts (the default) using an
archive tab, resulted in failure to correctly add a metric
to a chart.  The metric was "kernel.all.load[1 minute]" which
requires an indom lookup - this failed during the chart setup
process because of context switching at earlier points.