pmstat - fix bad error handling
When pmLookupName() fails for one or more metric names, the error
status is:
(a) different to that described in the man page and
(b) ambiguous.

This commit fixes (b) so pmstat does the extra work needed
to decode the error correctly for each metric than cannot be
looked up via pmLookupName().

Consideration of (a) is TBD.