Add hinv.cpu.model_name to counter the hinv.cpu.model scourge
For some time it turns out we've not had the most useful model
name being reported for each cpu in the hardware inventory.  It
turns out, because the order in which model and model_name are
presented/parsed, we end up using the less useful one on x86_64
(and IA64, and probably one or two other spots).  This adds in
a hinv.cpu.model_name metric to address this issue, and makes a
more coherent fallback strategy from one to the other (so we'll
always give a name if we can - /proc/cpuinfo contents are arch-
dependent and may have 0/1/2 of these fields).