Split PCP_TMP_DIR setting into its two distinct use-cases
Folks have complained that we are creating world-writable-sticky
bits in somewhat unorthadox locations, and question the need for
PCP to be doing that, arguing it opens up unnecessarily potential
exposure to security issues via attacks from hostile local users.

Auditing all of the current uses of PCP_TMP_DIR, shows that there
are two categories - PCP servers (daemons usually running as "pcp"
user, or "root" in the past) use this path to read/write files in
a way different programs can communicate (e.g. pmdammv, the status
files for pmie/pmlogger, etc) and for the creation of tmpfile(3)'s
that are short-lived.

For the former case, we do not strictly speaking need to use world
writable sticky directories - we could be using paths owned by pcp
(or just readable by pcp, in some cases).  For the latter case we
do need 'em, but we could be using the regular system locations for
temporary files.  There are many more cases of the former than the
latter usage.

This commit introduces a PCP_TMPFILE_DIR variable (for the latter),
usually /var/tmp - and keeps the existing PCP_TMP_DIR for the other
case (PCP daemons), usually /var/lib/pcp/tmp still.  PCP_TMP_DIR is
no longer installed world-writable-sticky-bit-set, rather it is now
a regular directory that only the pcp user/group can modify.

Slightly awkward in this scheme are pmdammv and pmdabash (and likely
other PMDAs in the future) - which need to share files between "pcp"
user daemons and arbitrary other user accounts.  They'll continue to
use $PCP_TMP_DIR/mmv, for example, but we do not enforce an initial
setup for this directory during the package installation anymore.
In other words, the user must decide on appropriate security setting
for their local site.  Any existing installation will continue on as
before, as will a manual PMDA Install.  However, the opportunity now
exists for a finer-grained scheme (e.g. specific-user-account owner,
pcp-group readable, and not world-writable-with-sticky-bit-set works
just fine).