Remove non-deterministic output during pcp (pmcd/pmlogger) restart
Dave reported numerous QA failures as a result of this issue, and I
too am seeing some.  A workaround existed in qa/187 so seems others
also have hit this.  There's a race condition in the restart script
for pcp which calls the pmcd and pmlogger variants which can result
in pmcd stopping but not yet pmlogger, before the pmlogger start is
performed.  This results in the get_running_pid check there finding
a pmlogger and reporting that its waiting, but there's no guarantee
being provided that the primary logger has reacted to the fact that
pmcd stopped before we get to that point.  Hence, a spurious output
message can occur.  Resolve this by making the scripts honour the
"quietly" first parameter in this section of the script (note that
a regular "start" (not "restart") will still report a need to delay
as a result of pmlogger presence.