Tackle further qa fallout from recent hostname updates
One other side-effect of no longer associating hostnames
with anything related to DNS is that the location of the
pmlogger archives for a host may have moved.  This causes
a handful of other tests to fail in non-obvious ways on a
test host that contains dots in its (real) hostname.