Remove PATH dependence in the collect2pcp tests
The collectl2pcp tests included a full path to the collectl2pcp
binary used in their output, which usually is just fine.  I've
come across a host today where /bin is a symlink to /usr/bin so
the PATH evaluation order then determines the output produced
by the tests - murphys law kicks in - and tests fail.

Resolve this by putting the path used in a seq.full file instead.