Solaris packaging fixes to install a pcp user/group correctly
Several changes relating to installing the pcp user and group
account as part of package installation.  Firstly, these need
to be done in a preinstall script, not postinstall, so we can
get file permissions right during installation.  Next, theres
no -r option to groupadd, nor -M/-r option to useradd.

Also add NeedRebuild logic into the postinstall as is done on
other platforms to ensure a namespace update on upgrade.