Shared library symbol versioning for libpcp_import
Addition of exports file, makefile tweaks, and also found
a handful of internal routines being made unintentionally
available (underscore prefixed).  These are never useful
to anyone outside the library so mark them hidden if gcc
is being used.

Converted the makefile to using DOMAIN_MAKERULE as well,
looks like it was overlooked in the last sweep.