Shared library symbol versioning for libpcp_trace
Add a library symbol version file for libpcp_trace.  Stop
installing the hash API header, this is deeply internal,
and nothing ever needs it (other than pmdatrace and lib).
Had to keep the dev header though as the QA tests use it.

Made a couple of unintentionally exposed symbols hidden.

Fix comment in trace.h about where demo programs can be
found, and add a note about the programmers guide too.

Remove an old Mac OS X hack in the trace PMDA makefile as
we nowadays (for quite some time) do not get __htonll in
both libpcp and libpcp_trace.

Also fixed the Fortran call interface for pmtraceobs_ as
it was using the AT&T Unix interface and not the version
supported by Linux/gcc.