Remove dependence on an unmaintainable internal libpcp symbol
In two places we had leaked out a dependence on a symbol that
definitely should have been internal to libpcp - __htonll.  We
simply cannot maintain symbols named like that in a shlib that
could be used by arbitrary applications, its a timebomb and it
will be awful for someone to have to debug/workaround a symbol
collision on that bad boy.

This removes those dependencies so we can properly internalise
this symbol like nature intended, in a future commit.