Remove endian macro/routine abuse in pmcd client handshake
Simplify the pmcd code which endian swaps the pduinfo word,
which is exchanged as part of an initial client connection.
There's no reason to expose this symbol in libpcp for just
this questionable use - its just a htonl-obfuscator.

This is the last external dependence on the endian code in
libpcp, so in a future commit we'll be able to internalise
this code and remove its unintended API/ABI footprint.