Shared library symbol versioning and hidden attributes for libpcp
Add a shared library symbol version file for libpcp, the last
of the PCP libraries to get this treatment and most difficult
of course.  As was done for the other libraries, this adds in
use of gcc features that allow any non-static library-internal
code and data to be marked as such as well.  Thus we now have
clear API and ABI definitons via the various headers and also
the src/libpcp/src/exports file.

There were three main areas that were more challenging to deal
with in libpcp, relative to the other libraries - they were the
endian-related code, derived metrics, and threading support.

For the first two, we were exporting symbols that weren't part
of the intended ABI (alot of ntoh/hton-named extensions and a
series of __dm* functions).  Code has been rearranged so we no
longer export these deeply-internal interfaces and so that we
do not have to keep them forever in the ABI definition.

For the latter (but also in the endian case) we were defining
some symbols conditionally, depending on threading support and
also thread-debugging support - problematic when attempting to
define a fixed set of interfaces.  This has been resolved, and
at the same time the code that was inline in the impl.h within
the PM_LOCK/PM_UNLOCK macros has been moved into the library,
improving its readability no end.  Unconditional debug messages
from those two macros have also been shifted to being guarded
by pmDebug settings (qa/src/exerlock.c is updated as a result).

As part of testing this work, the non-threaded libpcp build was
made to work once more - had to tweak check-statics to do so,
and resolve some warnings here and there in libpcp.

At the end of the day, new symbols added to libpcp (after the
next release; namely pcp-3.8.9) which we wish to make visible
to tools, need to be added to the exports file along with an
appropriate update to the version number there.  Unless such a
change is made, the symbol will not be exported.  Similarly,
any new internal symbol we add, that cannot be marked static,
should be marked as hidden (see _PCP_HIDDEN in internal.h).
This apparently allows the compiler to deal with it optimally.
This annotation only needs to be made within the header file.