Resolve a conflict between Avahi and pmcd with multiple ports
An intermittent sigsegv was picked up by qa/283 during recent
testing.  The cause of the problem is incorrect use of the
avahi-client API by libpcp when advertising a service on more
than one port.

Since Avahi requires the services to be committed together in
one service entry group, the Avahi implementation maintains a
table of active services. When a new service is added the table
is updated and the avahi-client is stopped while the new set of
services is recreated. The client is then restarted. A similar
process occurs when a service is removed.

Some re-factoring of code and data was also done which reflects
the new model for which there is only one avahi-client.