pmlogextract - dynamic indom botch with -S and -T
Reported by Tom Yearke.

When logging a dynamic indom, the timestamps are used to synchronize
the pmResults with the indom as it changes over time.

The logic was flawed, but requires a corner case to trigger the bug,
namely:
1. must have -S on the command line
2. must have 2 or more indom records in the input archive before the
   -S time
3. must have a pmResult after the -S time that correctly matches an
   indom record and then one or more additional (consecutive) pmResult
   records that use the _same_ indom state

When this happens, bogus indom records (from before the -S time
appear in the output archive with modified timestamps to match the
additional (consecutive) pmResult records.

Fixed by redoing the timestamp accounting and checking as indom
records are written to the output archive.